requirements.txt
openai
python-dotenv
pydantic
colorama
streamlit
